@@ -6,21 +6,23 @@ import BreedDetail from '../../Components/BreedDetail/BreedDetail.jsx';
 import CreateBreed from '../../Components/CreateBreed/CreateBreed';
 import Favorites from '../../Components/Favorites/Favorites';
 import Footer from '../../Components/Footer/Footer';
-import { Route } from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
 import './AppRouter.css';
 
 function AppRouter() {
   return (
     <div className='page-container'>
       <div className='content-wrap'>
-        <Route exact path='/' component={LandingPage} />
-        <Route path='/home' component={Menu} />
-        <Route exact path='/home' component={Breeds} />
-        <Route exact path='/home/:id' component={BreedDetail} />
-        <Route exact path='/home/breed/favorites' component={Favorites} />
-        <Route exact path='/home/breed/create' component={CreateBreed} />
+        <Routes>
+          <Route path='/' element={<LandingPage />} />
+          <Route path='/home' element={<Breeds />} />
+          <Route path='/home' element={<Menu />} />
+          <Route path='/home/:id' element={<BreedDetail />} />
+          <Route path='/home/breed/favorites' element={<Favorites />} />
+          <Route path='/home/breed/create' element={<CreateBreed />} />
+          <Route path='*/home' element={<Footer />} />
+        </Routes>
       </div>
-      <Route path='/home' component={Footer} />
     </div>
   );
 }
